backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
et_xmlfile==2.0.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.9.1  # typed config via BaseSettings
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyJWT==2.8.0
pytz==2025.2
email-validator==2.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
psycopg2-binary>=2.9.9
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
passlib[bcrypt]==1.7.4
